[SDK-1078] Registry-based URI association

+++ BranchSDK-Samples/Windows/TestBed/Button.h
@@ -10,7 +10,7 @@ class Button :
 {
 public:
     Button(LPCWSTR text, int x, int y, int width, int height, HMENU menu, HWND parent = nullptr);
-    Button(LPCWSTR text, int x, int y, int width, int height, int menu, HWND parent = nullptr);
+    Button(LPCWSTR text, int x, int y, int width, int height, long long menu, HWND parent = nullptr);
 
     void setButtonPressCallback(const std::function<void()>& buttonPressCallback)
     {
